[nrfconnect] Repair Rotating Device ID UID parsing
Rotating Device Id Unique Id was parsing wrongly in nrfconnect's
factory data provider implementation.

There was also a problem, that testing Rotating Device Id UID
was not being provided by Zephyr's Kconfig.

Repaired the factory_data build for nRFconnect devices.

+++ src/platform/nrfconnect/FactoryDataProvider.cpp
@@ -81,7 +81,7 @@ CHIP_ERROR FactoryDataProvider<FlashFactoryData>::Init()
         return error;
     }
 
-    if (!ParseFactoryData(factoryData, factoryDataSize, &mFactoryData))
+    if (!ParseFactoryData(factoryData, static_cast<uint16_t>(factoryDataSize), &mFactoryData))
     {
         ChipLogError(DeviceLayer, "Failed to parse factory data");
         return CHIP_ERROR_PERSISTED_STORAGE_VALUE_NOT_FOUND;
